Fixed Glitch Climbing Walls
The Ground Check box was the same size as the players box causing the player to be able to keep jumping none stop well hugging the wall. I changed the ground check box to be slightly smaller fixing the issue as now the only place that it can interact with anything is the bottom aka the ground.