Multi Jumping Not Working
Tried to get not just double jumping but able to customize the amount of jumps and make the system work with any number. Had it working for a bit however wasn't up to what I wanted and now it doesn't work at all will need to go back to this.